New UI theme and animations

+++ src/main/java/org/saturnclient/ui2/anim/SlideX.java
@@ -24,7 +24,7 @@ public class SlideX extends Animation {
     }
 
     @Override
-    public void tick(float progress, Element element) {
+    public void tick(double progress, Element element) {
         element.x = elementX + (int) (offset * progress);
     }
